Database modification functions
Adding and removing based on id from the database for tag, patient, dentist, scan. without env vars
